+++ src/pages/Article/textData.ts
